requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
gradio==5.11.0
gradio_client==1.5.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
protobuf==5.29.3
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyMatting==1.1.13
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rembg==2.0.61
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.6
safehttpx==0.1.6
scikit-image==0.25.0
scipy==1.15.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tifffile==2024.12.12
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
websockets==14.1
yarl==1.18.3
